Podcast: match Stats tab to dashboard idiom; swap By country for GeoChart
- Wrap Stats panel in .podcast__tab-content--wide (1024px max-width) so it
  reads at the same width as the other tabs instead of stretching to fill.
- Replace By country bar list with @automattic/charts GeoChart in a new
  Locations widget, mirroring the Calypso Stats page's list-locations card.
  The country breakdown was a flat list of mostly-one-entry rows; a world
  map reads at a glance for any data shape.
- Rebrand the muted text color: the existing rules used
  `--wp-admin-theme-color-darker-20` which resolves to the admin theme
  color (default blue), making body copy look like a link. Switch to the
  dashboard convention (#1e1e1e text / #757575 muted / #f0f0f0 bar bg).
- Match .podcast__section-heading (24px / weight 400 / #1e1e1e) for the
  period heading so it stops sticking out vs Settings / Distribution.
